Tests/LibWeb/Text/expected/indexeddb-result-after-abort-in-upgradeneeded.txt

result type: IDBDatabase
db.name: test-result-after-abort
